added empty state

+++ app/src/main/java/com/example/foodwaste/model/FoodItem.kt
@@ -3,7 +3,7 @@ package com.example.foodwaste.model
 data class FoodItem(
     val name: String = "",
     val expirationDate: String = "",
-    val co2: Int = 0,
+    val co2: Double = 0.0,
     var isChecked: Boolean = false,
     var isShared: Boolean = false
 )
